Change the label text from "View my solution" to "View Solution"
This commit changes the text of the label from "View my solution" to
"View solution".

As disscussed in the issue, campers visiting profile pages of others,
saw "View my solution" and found this confusing.

Removing the "my" from the text makes it generic, to either user's own
page or other's pages.

Tested locally.