Reformat a11y resources guide and add a resource (#18547)
* Reformat a11y resources guide and add a resource

I added a resource to an accessible color palette builder, as well as did some reformatting.

Color tools are under their own heading for easy perusable. General section could be broken down as well, I believe, but for now I think it should be ok as-is. 

Since the order of these resources is not important, I turned the lists into `ul`s, which makes more sense as well as making this guide easier to maintain. Now if someone wants to change the order around or add or remove something, there is no need to go and change all of the list item numbers.

* Update index.md