Added usages (#20035)
* Added usages

Added usages in the style of other files in the directory.

* Complete file overhaul.

Added details on how all of the three different modes are used. Added commonly used arguments section with seven arguments and how to use them. Added an examples section with three examples and explanations of them. Added a link to the SS64 page. Also made sure that the formatting is more consistent with the rest of the directory.

* Basic editting.

Removed hyphens before user (group) and read/write options because `-` is not required in front of the user (group) option and has a specific meaning (to remove) for the read/write options.

* Additional editing and formatting.

* Reworded the sentence for using a combination of numeric permission arguments.
* Changed the sentence that introduces commonly used arguments to a subheading.