Add the text 'My Changes' to article (#29892)
* Add the text 'My Changes' to article

The changes are on books that would guide children who are learning to programme, and also those who are unsure of where to start. I believe books are a great place for kids to start with learning how to code. These books would inspire them and help them understand the whole concept of programming, and not just as a pass-time, but as an art which goes beyond just writing codes and more into what makes our world be.

* Added 'Here are books to get kids coding'

* Formatted links

* Fixed book formatting