Edit on few texts (#29621)
* Edit on few texts

The first sort() function title was initially in Chinese, translated to the actual name of the function. 
The asort() function title was initially in uppercase, might confuse reader whether uppercase or lowercase to be user, converted to lowercase.

* Add New Book

Added Learning PHP, MySQL & JavaScript: With jQuery, CSS & HTML5 (Learning PHP, MYSQL, Javascript, CSS & HTML5) to the list.