Changed link to Spanish version of User History (#20165)
I have changed the link to the user history, which was originally in English. I switched it to one in Spanish. I then changed any grammatical errors to make it more cohesive and understandable. I separated the sentences to allow for a more user-friendly page. It was clumped up together initially and I think this will help a bit better.