Modify the text (#25275)
If a number couldn't be expressed in p/q form where p,q are real numbers and q in non-zero, the number is said to be an irrational number .To simplify the fraction we should rationalize the denominator. Rationalizing the denominator can aid you to simplify certain equations. This could be done by the following listed methods -