docs: fix Spanish documentation typos (#37294)
* Fix numerical representation on the Spanish docs

In the Spanish language, a dot should be used as thousands separator,
instead of a comma, like it happens in English.

This solves the situation for any numerical representation found over
the Spanish docs, to make it look professionally appealing for Spanish-
speaking audiences.

* Fix link to Open Source contributing from Spanish docs

* Fix thousand separator for Spanish README

In Spanish, thousands separator must be represented with a dot, instead of a comma. Comma is used to separate decimal digits.